+++ app/src/main/java/seamuslowry/daytracker/MainActivity.kt
@@ -1,0 +1,31 @@
+package seamuslowry.daytracker
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.ui.Modifier
+import dagger.hilt.android.AndroidEntryPoint
+import seamuslowry.daytracker.ui.navigation.Navigation
+import seamuslowry.daytracker.ui.navigation.Screen
+import seamuslowry.daytracker.ui.theme.PainTrackerTheme
+
+@AndroidEntryPoint
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            PainTrackerTheme {
+                // A surface container using the 'background' color from the theme
+                Surface(
+                    modifier = Modifier.fillMaxSize(),
+                    color = MaterialTheme.colorScheme.background,
+                ) {
+                    Navigation(startDestination = Screen.Entry.identifier)
+                }
+            }
+        }
+    }
+}
