+++ app/src/main/java/seamuslowry/daytracker/PainTrackerApplication.kt
@@ -1,0 +1,7 @@
+package seamuslowry.daytracker
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class PainTrackerApplication : Application()
