+++ app/src/main/java/seamuslowry/daytracker/data/room/daos/ItemConfigurationDao.kt
@@ -1,0 +1,21 @@
+package seamuslowry.daytracker.data.room.daos
+
+import androidx.room.Dao
+import androidx.room.Delete
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import androidx.room.Query
+import kotlinx.coroutines.flow.Flow
+import seamuslowry.daytracker.models.ItemConfiguration
+
+@Dao
+interface ItemConfigurationDao {
+    @Query("select * from item_configuration")
+    fun getAll(): Flow<List<ItemConfiguration>>
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    suspend fun upsert(itemConfiguration: ItemConfiguration): Long
+
+    @Delete
+    suspend fun delete(itemConfiguration: ItemConfiguration)
+}
