+++ app/src/main/java/seamuslowry/daytracker/di/DbModule.kt
@@ -1,0 +1,38 @@
+package seamuslowry.daytracker.di
+
+import android.content.Context
+import androidx.room.Room
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
+import dagger.hilt.components.SingletonComponent
+import seamuslowry.daytracker.data.room.PainTrackerDatabase
+import seamuslowry.daytracker.data.room.daos.ItemConfigurationDao
+import seamuslowry.daytracker.data.room.daos.ItemDao
+import javax.inject.Singleton
+
+@InstallIn(SingletonComponent::class)
+@Module
+object DbModule {
+    @Provides
+    @Singleton
+    fun providePainTrackerDb(
+        @ApplicationContext context: Context,
+    ): PainTrackerDatabase = Room
+        .databaseBuilder(context, PainTrackerDatabase::class.java, "pain_tracker_database")
+        .fallbackToDestructiveMigration()
+        .build()
+
+    @Provides
+    @Singleton
+    fun provideItemDao(
+        db: PainTrackerDatabase,
+    ): ItemDao = db.itemDao()
+
+    @Provides
+    @Singleton
+    fun provideItemConfigurationDao(
+        db: PainTrackerDatabase,
+    ): ItemConfigurationDao = db.itemConfigurationDao()
+}
