+++ app/src/main/java/seamuslowry/daytracker/models/TrackingType.kt
@@ -1,0 +1,6 @@
+package seamuslowry.daytracker.models
+
+enum class TrackingType {
+    ONE_TO_TEN,
+    YES_NO,
+}
