+++ app/src/main/java/seamuslowry/daytracker/ui/screens/report/ReportScreen.kt
@@ -1,0 +1,9 @@
+package seamuslowry.daytracker.ui.screens.report
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun ReportScreen() {
+    Text(text = "report")
+}
