--- app/src/main/java/seamuslowry/paintracker/PainTrackerApplication.kt
@@ -1,7 +1,0 @@
-package seamuslowry.paintracker
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class PainTrackerApplication : Application()
