--- app/src/main/java/seamuslowry/paintracker/data/room/Converters.kt
@@ -1,16 +1,0 @@
-package seamuslowry.paintracker.data.room
-
-import androidx.room.TypeConverter
-import java.time.LocalDate
-
-class Converters {
-    @TypeConverter
-    fun epochDayToLocalDate(value: Long?): LocalDate? {
-        return value?.let { LocalDate.ofEpochDay(value) }
-    }
-
-    @TypeConverter
-    fun localDateToEpochDay(date: LocalDate?): Long? {
-        return date?.toEpochDay()
-    }
-}
