--- app/src/main/java/seamuslowry/paintracker/models/Item.kt
@@ -1,28 +1,0 @@
-package seamuslowry.paintracker.models
-
-import androidx.room.ColumnInfo
-import androidx.room.Entity
-import androidx.room.ForeignKey
-import androidx.room.ForeignKey.Companion.CASCADE
-import androidx.room.PrimaryKey
-import java.time.LocalDate
-
-@Entity(
-    tableName = "item",
-    foreignKeys = [
-        ForeignKey(
-            entity = ItemConfiguration::class,
-            parentColumns = ["id"],
-            childColumns = ["configuration"],
-            onDelete = CASCADE,
-        ),
-    ],
-)
-data class Item(
-    @PrimaryKey(autoGenerate = true)
-    val id: Long = 0,
-    val date: LocalDate,
-    @ColumnInfo(index = true)
-    val configuration: Long,
-    val value: Int? = null,
-)
