--- app/src/main/java/seamuslowry/paintracker/models/ItemConfiguration.kt
@@ -1,13 +1,0 @@
-package seamuslowry.paintracker.models
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity(tableName = "item_configuration")
-data class ItemConfiguration(
-    @PrimaryKey(autoGenerate = true)
-    val id: Long = 0,
-    val name: String = "",
-    val trackingType: TrackingType = TrackingType.ONE_TO_TEN,
-    val active: Boolean = true,
-)
