--- app/src/main/java/seamuslowry/paintracker/models/TrackingType.kt
@@ -1,6 +1,0 @@
-package seamuslowry.paintracker.models
-
-enum class TrackingType {
-    ONE_TO_TEN,
-    YES_NO,
-}
