--- app/src/main/java/seamuslowry/paintracker/ui/screens/report/ReportScreen.kt
@@ -1,9 +1,0 @@
-package seamuslowry.paintracker.ui.screens.report
-
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-
-@Composable
-fun ReportScreen() {
-    Text(text = "report")
-}
